Extract WeakRefTrie from OrderByCanonicalizer.ts
