+++ packages/client/src/observable/internal/OrderByCanonicalizer.ts
@@ -15,59 +15,9 @@
  */
 
 import type { ObjectTypeDefinition } from "@osdk/api";
-import { Trie } from "@wry/trie";
 import type { OrderBy } from "../ObservableClient.js";
 import type { Canonical } from "./Canonical.js";
-
-const defaultMakeData = () => Object.create(null);
-
-export class WeakRefTrie<X extends object> {
-  #finalizer = new FinalizationRegistry<
-    Array<string>
-  >((orderBy) => {
-    this.#trie.removeArray(
-      Object.entries(orderBy).flat(),
-    );
-  });
-
-  #trie: Trie<WeakRef<X>>;
-
-  constructor(makeData: (array: any[]) => X = defaultMakeData) {
-    this.#trie = new Trie<WeakRef<X>>(
-      false,
-      (array) => {
-        const data = makeData(array);
-        this.#finalizer.register(data, array);
-        return new WeakRef(data);
-      },
-    );
-  }
-
-  lookupArray<T extends IArguments | any[]>(array: T): X {
-    const maybe = this.#trie.lookupArray(array);
-    let ret = maybe.deref();
-    if (maybe && !ret) {
-      // in case finalizer hasn't run
-      this.#trie.removeArray(array);
-      ret = this.#trie.lookupArray(array).deref();
-    }
-    return ret!;
-  }
-
-  peekArray<T extends IArguments | any[]>(array: T): X | undefined {
-    const maybe = this.#trie.peekArray(array);
-    const ret = maybe?.deref();
-    if (maybe && !ret) {
-      // in case finalizer hasn't run
-      this.#trie.removeArray(array);
-    }
-    return ret;
-  }
-
-  removeArray<T extends IArguments | any[]>(array: T): X | undefined {
-    return this.#trie.removeArray(array)?.deref();
-  }
-}
+import { WeakRefTrie } from "./WeakRefTrie.js";
 
 export class OrderByCanonicalizer {
   #trie = new WeakRefTrie(
